+++ psftp.c
@@ -41,7 +41,7 @@ bool sent_eof = false;
  * Seat vtable.
  */
 
-static int psftp_output(Seat *, bool is_stderr, const void *, int);
+static size_t psftp_output(Seat *, bool is_stderr, const void *, size_t);
 static bool psftp_eof(Seat *);
 
 static const SeatVtable psftp_seat_vt = {
@@ -2446,7 +2446,8 @@ void agent_schedule_callback(void (*callback)(void *, void *, int),
  */
 
 static bufchain received_data;
-static int psftp_output(Seat *seat, bool is_stderr, const void *data, int len)
+static size_t psftp_output(
+    Seat *seat, bool is_stderr, const void *data, size_t len)
 {
     /*
      * stderr data is just spouted to local stderr and otherwise
@@ -2477,7 +2478,7 @@ static bool psftp_eof(Seat *seat)
     return false;
 }
 
-bool sftp_recvdata(char *buf, int len)
+bool sftp_recvdata(char *buf, size_t len)
 {
     while (len > 0) {
         while (bufchain_size(&received_data) == 0) {
@@ -2486,19 +2487,19 @@ bool sftp_recvdata(char *buf, int len)
                 return false;          /* doom */
         }
 
-        int got = bufchain_fetch_consume_up_to(&received_data, buf, len);
+        size_t got = bufchain_fetch_consume_up_to(&received_data, buf, len);
         buf += got;
         len -= got;
     }
 
     return true;
 }
-bool sftp_senddata(const char *buf, int len)
+bool sftp_senddata(const char *buf, size_t len)
 {
     backend_send(backend, buf, len);
     return true;
 }
-int sftp_sendbuffer(void)
+size_t sftp_sendbuffer(void)
 {
     return backend_sendbuffer(backend);
 }
